+++ include/internal/catch_generators.hpp
@@ -14,7 +14,7 @@
 #include <memory>
 #include <vector>
 #include <cassert>
-
+#include <tuple>
 #include <utility>
 #include <exception>